[apidata] Move MemoryBuffers.h to ..
